# tests/fixtures/openapi.json
{
 "components": {
  "schemas": {
...
   "Card": {
    "description": "The snapshot fields, and ONLY the snapshot fields.\n\nThere is no ground_truth here and there is nowhere for one to appear: the\nmodel is closed over eight named fields, so even if a future query started\nselecting `*` from a table that had grown an answer column, this would refuse\nto carry it. §1 puts the reason plainly -- leak_check guards the line that a\ncard's answer must not be reachable, and an endpoint anyone with the port can\ncall is exactly where that line would break.",
    "properties": {
     "card_id": {
      "title": "Card Id",
      "type": "string"
     },
     "class": {
      "title": "Class",
      "type": "string"
     },
     "difficulty": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Difficulty"
     },
     "evidence_ok": {
      "title": "Evidence Ok",
      "type": "boolean"
     },
     "in_stock": {
      "title": "In Stock",
      "type": "boolean"
     },
     "primitive": {
      "title": "Primitive",
      "type": "string"
     },
     "snapshot_at": {
      "format": "date-time",
      "title": "Snapshot At",
      "type": "string"
     },
     "target": {
      "title": "Target",
      "type": "string"
     }
    },
    "required": [
     "card_id",
     "class",
     "target",
     "primitive",
     "difficulty",
     "in_stock",
     "evidence_ok",
     "snapshot_at"
    ],
    "title": "Card",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Page_Run_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/Run"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     }
    },
    "required": [
     "items",
     "next_cursor"
    ],
    "title": "Page[Run]",
    "type": "object"
   },
...
   "RunRef": {
    "properties": {
     "poll": {
      "title": "Poll",
      "type": "string"
     },
     "run_id": {
      "format": "uuid",
      "title": "Run Id",
      "type": "string"
     },
     "status": {
      "enum": [
       "queued",
       "running",
       "succeeded",
       "failed"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "run_id",
     "status",
     "poll"
    ],
    "title": "RunRef",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "rca-agent-bench evaluation service",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/cards": {
   "get": {
    "operationId": "list_cards_cards_get",
    "parameters": [
     {
      "in": "query",
      "name": "cardset",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cardset"
      }
     },
     {
      "in": "query",
      "name": "in_stock",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "In Stock"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Card"
         },
         "title": "Response List Cards Cards Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Cards"
   }
  },
  "/cards/{card_id}": {
   "get": {
    "operationId": "get_card_cards__card_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "card_id",
      "required": true,
      "schema": {
       "title": "Card Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Card"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Card"
   }
  },
...
  "/runs": {
   "get": {
    "operationId": "list_runs_runs_get",
    "parameters": [
     {
      "in": "query",
      "name": "card",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Card"
      }
     },
     {
      "in": "query",
      "name": "arm",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Arm"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "model",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Model"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 200,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_Run_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Runs"
   },
   "post": {
    "description": "Accept a run, or refuse it before it can cost anything.\n\nThe four gates run in this order because each is cheaper and more certain than\nthe next, and because §4 is explicit that the expensive checks must not be\nwhat protects the budget:\n\n  1. card_not_found (404) -- a card not in the snapshot, or not in stock.\n  2. evidence_leak / evidence_missing (422) -- leak_check.check_card on the\n     pack; a LeakError is the first, an absent or unreadable pack the second\n     (see the error-code table in models.py). run_card runs\n     this again internally before the first token; this outer one exists\n     because the service adds \"automatic, batched, unattended\" to the picture,\n     and in that setting a broken pack should never enter a system that will\n     retry it (§4).\n  3. the idempotency key -- looked up BEFORE the budget gate. A replay does\n     not spend anything: it hands back a run that already exists. Making it\n     pass the budget check would mean that the moment the daily budget runs\n     out, a retrying client stops being able to find out its own run_id --\n     the breaker would start withholding information instead of withholding\n     money. The gate belongs in front of work that will newly charge.\n  4. budget_exceeded (429) -- the daily breaker, on new runs only. The rules\n     arm passes through it too even though it costs nothing: one rule that is\n     always true beats two rules that are usually true, and an arm-dependent\n     breaker is exactly the sort of thing that is correct until someone adds a\n     fourth arm.",
    "operationId": "create_run_runs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "202": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunRef"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Run"
   }
  },
  "/runs/{run_id}": {
   "get": {
    "operationId": "get_run_runs__run_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Run"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run"
   }
  },
  "/runs/{run_id}/trace": {
   "get": {
    "operationId": "get_run_trace_runs__run_id__trace_get",
    "parameters": [
     {
      "in": "path",
      "name": "run_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Run Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunTrace"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Run Trace"
   }
  },
  "/summary": {
   "get": {
    "description": "The four-metric table, aggregated by compare_arms.metrics() (决策 037).",
    "operationId": "get_summary_summary_get",
    "parameters": [
     {
      "in": "query",
      "name": "cardset",
      "required": true,
      "schema": {
       "title": "Cardset",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "arms",
      "required": true,
      "schema": {
       "title": "Arms",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "pick",
      "required": false,
      "schema": {
       "default": "latest",
       "title": "Pick",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "runs",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Runs"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Summary"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Summary"
   }
  }
 }
}